backend_fastapi/requirements.txt
# 论文评价分析系统 - Python依赖包

# FastAPI核心框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# 文件处理和上传
python-multipart>=0.0.6
aiofiles>=23.0.0
python-docx>=0.8.11

# 系统监控和性能
psutil>=5.9.0

# 环境配置
python-dotenv>=1.0.0

# HTTP请求
requests>=2.31.0

# AI模型接口
openai>=1.3.0
google-generativeai>=0.3.0
tiktoken>=0.5.1

# 数据处理
pandas>=2.1.3
numpy>=1.24.3

# 文档处理（继承原项目）
mammoth>=1.6.0
lxml>=4.9.3
beautifulsoup4>=4.12.2
Pillow>=10.0.1

# 可视化（可选，用于数据分析）
plotly>=5.17.0

# 安全相关（可选）
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 测试框架（开发依赖）
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
